[MNG-7324] split config/domain/diff/report to separate xml.* packages
