+++ maven-core/src/main/java/org/apache/maven/caching/CacheControllerImpl.java
@@ -31,18 +31,18 @@ import org.apache.maven.artifact.handler.ArtifactHandler;
 import org.apache.maven.artifact.handler.manager.ArtifactHandlerManager;
 import org.apache.maven.caching.checksum.KeyUtils;
 import org.apache.maven.caching.checksum.MavenProjectInput;
-import org.apache.maven.caching.domain.Scm;
+import org.apache.maven.caching.xml.config.PropertyNameType;
+import org.apache.maven.caching.xml.config.TrackedPropertyType;
+import org.apache.maven.caching.xml.diff.BuildDiffType;
+import org.apache.maven.caching.xml.domain.ArtifactType;
+import org.apache.maven.caching.xml.domain.CompletedExecutionType;
+import org.apache.maven.caching.xml.domain.DigestItemType;
+import org.apache.maven.caching.xml.domain.ProjectsInputInfoType;
+import org.apache.maven.caching.xml.domain.Scm;
+import org.apache.maven.caching.xml.report.CacheReportType;
+import org.apache.maven.caching.xml.report.ProjectReportType;
 import org.apache.maven.caching.hash.HashAlgorithm;
 import org.apache.maven.caching.hash.HashFactory;
-import org.apache.maven.caching.domain.ArtifactType;
-import org.apache.maven.caching.domain.BuildDiffType;
-import org.apache.maven.caching.domain.CacheReportType;
-import org.apache.maven.caching.domain.CompletedExecutionType;
-import org.apache.maven.caching.domain.DigestItemType;
-import org.apache.maven.caching.domain.ProjectReportType;
-import org.apache.maven.caching.domain.ProjectsInputInfoType;
-import org.apache.maven.caching.domain.PropertyNameType;
-import org.apache.maven.caching.domain.TrackedPropertyType;
 import org.apache.maven.caching.xml.BuildInfo;
 import org.apache.maven.caching.xml.CacheConfig;
 import org.apache.maven.caching.xml.CacheSource;
