+++ maven-core/src/main/java/org/apache/maven/caching/HttpRepositoryImpl.java
@@ -33,10 +33,10 @@ import org.apache.http.entity.InputStreamEntity;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.caching.checksum.MavenProjectInput;
-import org.apache.maven.caching.domain.ArtifactType;
-import org.apache.maven.caching.domain.BuildInfoType;
-import org.apache.maven.caching.domain.CacheReportType;
-import org.apache.maven.caching.domain.ProjectReportType;
+import org.apache.maven.caching.xml.domain.ArtifactType;
+import org.apache.maven.caching.xml.domain.BuildInfoType;
+import org.apache.maven.caching.xml.report.CacheReportType;
+import org.apache.maven.caching.xml.report.ProjectReportType;
 import org.apache.maven.caching.xml.BuildInfo;
 import org.apache.maven.caching.xml.CacheConfig;
 import org.apache.maven.caching.xml.CacheSource;
