+++ maven-core/src/main/java/org/apache/maven/caching/RemoteArtifactsRepository.java
@@ -20,8 +20,8 @@ package org.apache.maven.caching;
  */
 
 import com.google.common.base.Optional;
-import org.apache.maven.caching.domain.ArtifactType;
 import org.apache.maven.caching.xml.BuildInfo;
+import org.apache.maven.caching.xml.domain.ArtifactType;
 import org.apache.maven.project.MavenProject;
 
 import java.io.IOException;
