+++ maven-core/src/main/java/org/apache/maven/caching/xml/CacheConfigImpl.java
@@ -26,26 +26,26 @@ import org.apache.maven.caching.DefaultPluginScanConfig;
 import org.apache.maven.caching.PluginScanConfig;
 import org.apache.maven.caching.PluginScanConfigImpl;
 import org.apache.maven.caching.checksum.MultimoduleDiscoveryStrategy;
-import org.apache.maven.caching.domain.AttachedOutputs;
-import org.apache.maven.caching.domain.Exclude;
-import org.apache.maven.caching.domain.Include;
-import org.apache.maven.caching.domain.Local;
-import org.apache.maven.caching.domain.ProjectDiscoveryStrategy;
-import org.apache.maven.caching.domain.Remote;
 import org.apache.maven.caching.hash.HashFactory;
-import org.apache.maven.caching.domain.CacheType;
-import org.apache.maven.caching.domain.ConfigurationType;
-import org.apache.maven.caching.domain.CoordinatesBaseType;
-import org.apache.maven.caching.domain.ExecutablesType;
-import org.apache.maven.caching.domain.ExecutionConfigurationScanType;
-import org.apache.maven.caching.domain.ExecutionControlType;
-import org.apache.maven.caching.domain.ExecutionIdsListType;
-import org.apache.maven.caching.domain.GoalReconciliationType;
-import org.apache.maven.caching.domain.GoalsListType;
-import org.apache.maven.caching.domain.PluginConfigurationScanType;
-import org.apache.maven.caching.domain.PluginSetType;
-import org.apache.maven.caching.domain.PropertyNameType;
-import org.apache.maven.caching.domain.TrackedPropertyType;
+import org.apache.maven.caching.xml.config.AttachedOutputs;
+import org.apache.maven.caching.xml.config.Exclude;
+import org.apache.maven.caching.xml.config.Include;
+import org.apache.maven.caching.xml.config.Local;
+import org.apache.maven.caching.xml.config.ProjectDiscoveryStrategy;
+import org.apache.maven.caching.xml.config.Remote;
+import org.apache.maven.caching.xml.config.CacheType;
+import org.apache.maven.caching.xml.config.ConfigurationType;
+import org.apache.maven.caching.xml.config.CoordinatesBaseType;
+import org.apache.maven.caching.xml.config.ExecutablesType;
+import org.apache.maven.caching.xml.config.ExecutionConfigurationScanType;
+import org.apache.maven.caching.xml.config.ExecutionControlType;
+import org.apache.maven.caching.xml.config.ExecutionIdsListType;
+import org.apache.maven.caching.xml.config.GoalReconciliationType;
+import org.apache.maven.caching.xml.config.GoalsListType;
+import org.apache.maven.caching.xml.config.PluginConfigurationScanType;
+import org.apache.maven.caching.xml.config.PluginSetType;
+import org.apache.maven.caching.xml.config.PropertyNameType;
+import org.apache.maven.caching.xml.config.TrackedPropertyType;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.model.Plugin;
 import org.apache.maven.model.PluginExecution;
